site.css; made posts edit private to that user; bids not visibile to other bidders,but can edit own bid.

+++ BidCraft.web/Models/BidIndexVM.cs
@@ -7,12 +7,9 @@ namespace BidCraft.web.Models
 {
     public class BidIndexVM
     {
-        public int Amount { get; set; }
+        public decimal Amount { get; set; }
         public DateTime ProjectFinishByFinishDate { get; set; }
-        public int PostId
-        {
-            get; set;
-        }
+        public int PostId { get; set; }
 
     }
 }
